+++ frontend/src/layouts/StandardPage.js
@@ -1,20 +1,21 @@
-
 import React from 'react'
 import { Container } from 'react-bulma-components'
 import TopNavbar from '../components/TopNavbar'
 
- const StandardPage = ({children}) => {
+const StandardPage = ({ children }) => {
   return (
     <>
-      <Container max breakpoint="desktop" renderAs="header">
+      <header>
         <TopNavbar />
+      </header>
+      <Container max breakpoint="desktop" className="main" renderAs="main">
+        {children}
       </Container>
-      <Container max breakpoint="desktop" renderAs="main">
-          {children}
-      </Container>
+      <footer className="footer">
+        <Container textAlign="center">&copy; TiMalo — 2022</Container>
+      </footer>
     </>
   )
 }
- 
 
 export default StandardPage