+++ frontend/src/pages/LoginPage.js
@@ -1,33 +1,64 @@
-import React from 'react'
+import React, { useState, useEffect } from 'react'
 import LoginForm from '../components/LoginForm'
 import StandardPage from '../layouts/StandardPage'
-import { useAuth } from '../components/AuthProvider'
+import { Box, Button, Heading, Notification } from 'react-bulma-components'
+import { useZakari } from '../components/ZakProvider'
+import { useNavigate } from 'react-router-dom'
 
 const LoginPage = () => {
-    const auth = useAuth()
+  const [message, setMessage] = useState('')
+  const auth = useZakari()
+  const navigate = useNavigate()
 
-    const onSubmit = async ({email, password, setLoading}) =>  {
-        try {
-            setLoading(true)
-            const { error } = await auth.signIn({
-              email: email,
-              password: password,
-            })
-      
-            if (error) throw error
-            alert('Check your email for the login link!')
-          } catch (error) {
-            alert(error.error_description || error.message)
-          } finally {
-            setLoading(false)
-          }
-      
+  useEffect(() => {
+    if (auth.user !== null) {
+      navigate('/')
     }
+  })
+
+  const onSubmit = async ({ email, password, setLoading }) => {
+    console.log({ email, password, setLoading })
+    try {
+      setLoading(true)
+      setMessage('  ')
+      auth.signIn(email, password).then(
+        () => {
+          navigate('/')
+        },
+        (reason) => {
+          const code = reason?.code || 500
+          const msg =
+            code === 500
+              ? 'Erreur inconnue, veuillez essayer ultérieurement'
+              : 'Identifiant inconnu ou mot de passe incorrect'
+          setMessage(msg)
+        }
+      )
+    } catch (error) {
+      console.log(error)
+    } finally {
+      setLoading(false)
+    }
+  }
   return (
     <StandardPage>
-        <LoginForm buttonLabel='Login' onSubmit={onSubmit} />
+      <Box className="login_form">
+        <Heading size={2} renderAs="h1">
+          Log in
+        </Heading>
+        {message.length > 0 ? (
+          <Notification color="danger">
+            {message}
+            <Button remove onClick={() => setMessage('')} />
+          </Notification>
+        ) : (
+          ''
+        )}
+
+        <LoginForm buttonLabel="Login" onSubmit={onSubmit} />
+      </Box>
     </StandardPage>
   )
-}   
+}
 
 export default LoginPage